+++ app/Http/Controllers/FeedbackController.php
@@ -11,20 +11,36 @@ use App\Models\Feedback;
 
 class FeedbackController extends Controller
 {
-    public function store(FeedbackRequest $request)
+    public function store(FeedbackRequest $request, $accountId)
     {
+        $account = Account::find($request->from_account_id);
+        if (empty($account)) {
+            return $this->commonResponse([], "Account does not exist!", 404);
+        };
         $feedback = Feedback::create([
             'booking_id' => $request->booking_id,
-            'from_type' => $request->role_id,
-            'from_account_id' => $request->account_id,
+            'from_type' => $account->role_id,
+            'from_account_id' => $request->from_account_id,
+            'to_account_id' => $accountId,
             'content' => $request->content,
         ]);
 
         //send notifications for brand/influencer 
 
         $booking = Booking::find($feedback->booking_id);
+            if (empty($booking)) {
+                return $this->commonResponse([], "Booking does not exist!", 404);
+            };
+
         $influencer = Account::find($booking->influencer_id);
+            if (empty($booking)) {
+                return $this->commonResponse([], "Influencer does not exist!", 404);
+            };
+
         $campaign = Campaign::find($booking->campaign_id);
+            if (empty($booking)) {
+                return $this->commonResponse([], "Campaign does not exist!", 404);
+            };
 
         if ($feedback->from_type == Account::ROLE_BRAND) {
             $influencerNotifyContent = 'Hi @' . $influencer->username . ', you have just received a feedback from brand for booking with the ID #' . $booking->id . ' state now!';
@@ -40,7 +56,10 @@ class FeedbackController extends Controller
 
     public function edit(FeedbackRequest $request, $id)
     {
-        $feedback = Feedback::findOrFail($id);
+        $feedback = Feedback::find($id);
+        if (empty($feedback)) {
+            return $this->commonResponse([], "Feedback does not exist!", 404);
+        }
         $feedback->update([
             'content' => $request->content,
         ]);
@@ -50,7 +69,10 @@ class FeedbackController extends Controller
 
     public function destroy($id)
     {
-        $feedback = Feedback::findOrFail($id);
+        $feedback = Feedback::find($id);
+        if (empty($feedback)) {
+            return $this->commonResponse([], "Feedback does not exist!", 404);
+        }
         $feedback->delete();
 
         return $this->commonResponse([], "Feedback has deleted successfully.");
